Anchored patterns and moved a function to a test file.
